tweetapi/requirements.txt
fastapi
uvicorn
pydantic
psycopg2
SQLAlchemy
Jinja2
grpcio
grpcio-tools
kafka-python
redis
redis-server
